# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API UKM Politeknik",
        "description": "API UKM Politeknik Negeri Lampung. Response dapat berubah sewaktu waktu menyesuaikan perbaikan",
        "contact": {
            "email": "leopradana013@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/private/pendaftaran-anggota/list": {
            "get": {
                "tags": [
                    "Pendaftaran Anggota"
                ],
                "summary": "Daftar Pendaftaran Anggota",
                "operationId": "29a3a7cb1de740751ad16e4a4c5ddbec",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/private/pendaftaran-anggota/{id_pendaftaran_anggota}": {
            "delete": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Menghapus pendaftaran anggota",
                "description": "Menghapus informasi pendaftaran anggota",
                "operationId": "31c20c2d6d78330b7f7d876886cfa194",
                "parameters": [
                    {
                        "name": "id_pendaftaran_anggota",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/private/ukm": {
            "get": {
                "tags": [
                    "UKM"
                ],
                "summary": "Detail UKM",
                "description": "Informasi lengkap data UKM",
                "operationId": "63fcc3e0222555b7aa56aefc002c198d",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "UKM"
                ],
                "summary": "Perbarui UKM",
                "description": "Perbarui informasi UKM",
                "operationId": "87c566e2be701c384e2fc2a8214bbdc8",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "foto_ukm",
                                    "nama",
                                    "singkatan_ukm",
                                    "tentang_kami",
                                    "visi",
                                    "misi",
                                    "tujuan",
                                    "facebook",
                                    "instagram",
                                    "twitter",
                                    "whatsapp"
                                ],
                                "properties": {
                                    "foto_ukm": {
                                        "description": "Logo UKM",
                                        "type": "file"
                                    },
                                    "nama": {
                                        "description": "Nama UKM",
                                        "type": "string"
                                    },
                                    "singkatan_ukm": {
                                        "description": "Singkatan UKM",
                                        "type": "string"
                                    },
                                    "tentang_kami": {
                                        "description": "Tentang UKM",
                                        "type": "string"
                                    },
                                    "visi": {
                                        "description": "Visi UKM",
                                        "type": "string"
                                    },
                                    "misi": {
                                        "description": "Misi UKM",
                                        "type": "string"
                                    },
                                    "tujuan": {
                                        "description": "Tujuan UKM",
                                        "type": "string"
                                    },
                                    "facebook": {
                                        "description": "Link facebook UKM",
                                        "type": "string"
                                    },
                                    "instagram": {
                                        "description": "Link instagram UKM",
                                        "type": "string"
                                    },
                                    "twitter": {
                                        "description": "Link twitter UKM",
                                        "type": "string"
                                    },
                                    "whatsapp": {
                                        "description": "Nomor whatsapp UKM",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}